feat(nat): Implement "prefix collapsing" for stateless NAT
Implement a function to "collapse" prefixes: take a set of allowed
prefixes, and another set of exclusion prefixes to apply to the first
set, and produce a new set of allowed prefixes that are smaller segments
representing the same IP address range as the combination of the initial
two sets - but without the use of exclusion prefixes.

Implement a function to run this on all expose objects in a Peering. The
function is not called yet.

Add unit tests to validate the algorithm.

Signed-off-by: Quentin Monnet <qmo@qmon.net>