feat(nat): Simplify struct TrieValue
Instead of holding the full BTreeSet<Prefix> for the original prefixes
for the mapping, just store the prefix in use with its offset in the
list. This is more efficient in terms of memory usage, and saves us from
recomputing the offset of this prefix at runtime when computing the IP
address mapping.

Signed-off-by: Quentin Monnet <qmo@qmon.net>