fix(nat): Fix validation for one-direction NAT
This is similar to commit 7344ee06abbd ("fix(nat): Use ips if as_range
are empty for stateless src NAT 1st lookup"): when we only NAT on one
direction, then one of the two VPCs uses its "ips" prefix list as public
IPs (no NAT), and these are the prefixes we should compare public
addresses from as_range in other expose objects with.

In particular, we need to do this comparison when validating VpcManifest
objects: when checking for prefix overlap, we may have to check the
as_range blocks of an expose object with the ips blocks of another, if
the latter expose has no as_range.

Note that even in that case, we still check for overlap between the ips
lists of both expose objects, as we can't afford to have overlap for
them (ips lists of exposes in a same manifest).

Signed-off-by: Quentin Monnet <qmo@qmon.net>