feat(routing): add vni-based nexthop/path selector
If packets come with vni annotation, use only those next-hops that
use vxlan encapsulation for the indicated vni. This is needed to
disambiguate adress overlapping among VPCs when importing host
routes to a vrf. If, in that case, more than one next-hop exists,
ecmp among them. Moreover, if if the route is a host route, then
issue a warning since that means that the same host is reachable
over distinct leaves.

This patch won't have any effect if packets are not annotated.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>