feat(packet-processor): override ingress vrf
Override ingress vrf lookup if packet comes tagged with a vni.
This is the second approach to integrate nat/vpc-routing that
does not require vrf imports. We enable this here, because
the import approach does not work because frr seems not to
allow importing more than one host route from distinct vrfs.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>