fix(nat): Handle prefix size correctly
It looks like we've not been careful with prefix sizes. The method
prefix.size() would return the number of IP addresses available within
the prefix, as a u128. But for IPv6 prefix "::/0", the size is 2^128,
which is more than u128::MAX (2^128 - 1).

We also have another issue when adding prefixes offsets and sizes to
compare them with another value can overflow a u128, for example the
offset (2^64) and size (2^64) for "8000::/1".

The second case is easy to address, by moving around the members
involved in the comparison to avoid the addition and the overflow.

The first case is much more delicate, and requires the use of a
different type to hold the size of prefixes. This commit introduce the
enum PrefixSize, which can be a normal u128, a special Ipv6RootMax value
representing 2^128, and an Overflow value only used for operations. We
implement several traits for PrefixSize and u128: Add, Sub, Sum,
PartialEq, From/TryFrom. This helps use manipulate these objects
elsewhere in the code.

Even with this new type, some local adjustments are required to harden
the checks on prefix sizes, in order to avoid overflows or logic bugs.
We adjust the code at these locations to use PrefixSize and to run the
relevant checks accordingly. We also add some unit tests for PrefixSize.

Signed-off-by: Quentin Monnet <qmo@qmon.net>