feat(nat): Optimise stateless NAT lookup (many prefixes -> 2 IPs)
This commit changes the way we store information for stateless NAT
mapping, and the way we process to the lookup and mapping computation.

The former model used IpList and IpListSubset (an optimised version of
the former) objects. These objects represent the whole sets of allowed
original and target IP addresses for the VpcExpose, and we'd retrieve
the offset of the original IP within the first set, then look for the
corresponding IP in the second set.

This method works, but it's not great in terms of efficiency: it means
that we need to loop over all the prefixes each time we want an offset
(although the IpListSubset optimisation helped, here) and then again
when we want to map this offset to the target address. Not to mention
the memory cost of saving both lists of prefixes as a TrieValue, for
each possible prefix, both for source and destination NAT.

Let's rework the way we build the NAT tables. Instead of storing all
these prefix lists in TrieValue objects and to associate them to each
possible original prefix, we split the list of original prefixes into
chunks, each of which gets the exact chunk (from the target prefixes)
into which the addresses map, at the same offset.

Okay, this sounds hard to visualise without a picture. Here's the
legacy method, YAML-ish style:

  dst_nat table:
     - key: 1.0.0.0/16
       value:
         TrieValue:
          original prefixes (IpList 1):
            - 1.0.0.0/16
            - 2.0.0.0/16
          destination prefixes (IpList 2):
            - 10.1.0.0/17
            - 10.2.0.0/17
            - 10.3.0.0/16
     - key: 2.0.0.0/16
       value:
         TrieValue:
          original prefixes (IpList 1):
            - 1.0.0.0/16
            - 2.0.0.0/16
          destination prefixes (IpList 2):
            - 10.1.0.0/17
            - 10.2.0.0/17
            - 10.3.0.0/16

From there, if a packet has destination 2.0.2.5, we compute the offset
of the address within the first IpList (2^16 + 2*2^8 + 5), then look for
the address at the same offset in the second set, and obtain 10.3.2.5.

The IpListSubset optimisation shrinks one of the two IpList objects.
here's the existing representation of the table:

  dst_nat table:
     - key: 1.0.0.0/16
       value:
         TrieValue:
          original prefixes (IpListSubset):
            prefix: 1.0.0.0/16
            offset: 0
          destination prefixes (IpList):
            - 10.1.0.0/17
            - 10.2.0.0/17
            - 10.3.0.0/16
     - key: 2.0.0.0/16
       value:
         TrieValue:
          original prefixes (IpListSubset):
            prefix: 2.0.0.0/16
            offset: 2^16
          destination prefixes (IpList):
            - 10.1.0.0/17
            - 10.2.0.0/17
            - 10.3.0.0/16

This time, we only have one tuple (prefix, prefix offset) in the
IpListSubset, and not the full list of original prefixes. When we
receive 2.0.2.5, we compute the offset within the prefix (2*2^8 + 5) and
add the prefix offset (2^16), to obtain the same offset and then the
same mapping as in the first case.

Now, the new model in this commit is the following:

  dst_nat table:
     - key: 1.0.0.0
       value:
         original prefix end: 1.0.127.255
         target range start: 10.1.0.0
     - key: 1.0.128.0
         original prefix end: 1.0.255.255
         target range start: 10.2.0.0
     - key: 2.0.0.0
         original prefix end: 2.0.255.255
         target range start: 10.3.0.0

Note that we have more keys (the 1.0.0.0/16 prefix was split into two
chunks, each being associated to a /17 range from the target prefixes;
target prefixes can be split too, if necessary, so that we keep a 1:1
address mapping between the associated ranges). We have more entries in
the table, but they are much smaller, each containing 3 IP addresses
only (one for the key, two making the value).

These entries are stored in a BTreeMap. When our packet with destination
2.0.2.5 arrives, we look in the map for the closest lower (or equal) key
in the map corresponding to the entry. In our case, this is 2.0.0.0. The
lookup returns the full entry, so we have the original prefix start and
end addresses, and the target range start address. From here, we compute
the offset of the address within the range (2*2^8 + 5), and retrieve the
address at the same offset within the target range (10.3.2.5).

We don't need the original prefix end address to compute the offset.
However, we do need this end address to make sure that the entry we find
in the table is the right one: we don't want to find the closest lower
range start address if the address we're looking up is not part of this
range. This happens for source NAT, for example, when we search the
different relevant expose sub-tables for the range corresponding to the
packet source IP.

The most complex part of this commit is the change of code for building
the NAT rule entries, because we need to split the existing prefixes
into these ranges to preserve the 1:1 mapping between each pair of
ranges. We also have some sensitive updates in the structure of the NAT
rules, and the insert() and lookup() function for these tables. In the
core logic for the NAT code, it looks like the code for the mapping gets
bigger, but this is because the new functions replace all of the IpList
processing, that we don't use anymore. The IpList submodule will be
entirely removed in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>