feat(net): build packets with 'keep' flag set
Up until now, stages / NFs may filter out packets if marked
with a certain DoneReason. Enable 'keep' flag that was only
for tests so that packet iterators never filter packets,
allowing them all to reach a last statistics stage.
It is to be seen the performance penalty of this, but
this should not worsen the worst-case scenario in terms of
effort where all packets would be forwarded.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>