fix(nat): Make stateful NAT code direction-independent
In commit b12d185a5e87 ("feat(nat): Make NAT stage
direction-independant") we changed the way we're handling directions in
NAT stages. That commit mostly focused on stateful NAT, and the
sensitive code portions in the module for stateful NAT were simply
commented out, given that the code is not complete anyway.

Here we adjust the code for stateful NAT, on the same basis. We make it
handle both source and destination NAT, in a single pipeline stage. This
requires storing more information in the state tables, given that we
need addresses and port for both directions.

Signed-off-by: Quentin Monnet <qmo@qmon.net>