feat(mgmt): Merge contiguous ranges when building stateless NAT tables
When building stateless NAT tables, we turn our list of prefixes into
different ranges, and insert each of these ranges into the NAT tables.
We cannot easily merge and process contiguous CIDR prefixes; however,
IP ranges, being defined with just a start and an end address, are easy
to concatenate. Make sure to "merge" (concatenate) ranges from
contiguous prefixes when we process them to build the tables.

For an illustration, see the example from the unit test. We have:

    let prefixes_to_update = BTreeSet::from([
        "1.0.0.0/24".into(),
        "1.0.1.0/24".into(),
        "1.0.2.0/24".into(),
        "1.0.3.0/24".into(),
        "2.0.0.0/16".into(),
        "2.1.0.0/16".into(),
        "2.2.0.0/16".into(),
        "2.3.0.0/16".into(),
    ]);
    let prefixes_to_point_to = BTreeSet::from([
        "10.0.0.0/24".into(),
        "10.0.1.0/24".into(),
        "10.0.2.0/24".into(),
        "11.0.0.0/16".into(),
        "11.1.0.0/16".into(),
        "11.2.0.0/16".into(),
        "11.3.0.0/16".into(),
        "11.4.0.0/24".into(),
    ]);

Note how the first set has a last /24 prefix when the second switches to
a series of /16, breaking the "alignment" and causing more ranges to be
created. Converted into basic mapping ranges, this would produce:

    1.0.0.0-1.0.0.255     -> 10.0.0.0-10.0.0.255
    1.0.1.0-1.0.2.255     -> 10.0.1.0-10.0.1.255
    1.0.2.0-1.0.2.255     -> 10.0.2.0-10.0.2.255
    1.0.3.0-1.0.3.255     -> 11.0.0.0-11.0.0.255
    2.0.0.0-2.0.254.255   -> 11.0.1.0-11.0.255.255
    2.0.255.0-2.0.255.255 -> 11.1.0.0-11.1.0.255
    2.1.0.0-2.1.254.255   -> 11.1.1.0-11.1.255.255
    2.1.255.0-2.1.255.255 -> 11.2.0.0-11.2.0.255
    2.2.0.0-2.2.254.255   -> 11.2.1.0-11.2.255.255
    2.2.255.0-2.3.255.255 -> 11.3.0.0-11.3.0.255
    2.3.0.0-2.3.254.255   -> 11.3.1.0-11.3.255.255
    2.3.255.0-2.3.255.255 -> 11.4.0.0-11.4.0.255

But most of these ranges are contiguous, and can be merged to produce
fewer entries. The following mappings are equivalent:

    1.0.0.0-1.0.2.255    -> 10.0.0.0-10.0.2.255
    1.0.3.0-1.0.3.255    -> 11.0.0.0-11.0.0.255
    2.0.0.0-2.3.255.255  -> 11.0.1.0-11.4.0.255

This is what we implement in this commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>