fix(prefix): Fix Add implementation for PrefixSize + PrefixSize
We based our implementation on the addition of a PrefixSize and a u128,
making all other cases return a PrefixSize::Overflow, but the addition
between two PrefixSize objects needs to take another corner case into
account: for PrefixSize::U128(0) + PrefixSize::Ipv6MaxAddrs, the result
is not an overflow but PrefixSize::Ipv6MaxAddrs.

Found while working on fuzzing tests for another component.

Signed-off-by: Quentin Monnet <qmo@qmon.net>