test(mgmt): Add fuzzing tests (bolero) for stateless NAT's range_builder
Add fuzzing tests in which we generate two lists of prefixes, pick
addresses within the first list, and make sure that we can find IP
ranges within the generated NAT ranges from the first prefix list. Make
sure these NAT ranges have a total size that matches the size (number of
IP addresses covered) of the prefix list.

Generating the required prefix lists is non-trivial, given that each
list must only contain non-overlapping prefixes, and both lists must
have the same total size (total number of IP addresses covered by the
prefixes), witout necessarily containing prefixes of the same lengths.
Much of the code is dedicated to the generation of these lists.

This test led to the discovery of a bug on the implementation of the Add
trait for PrefixSize, which has been addressed in a previous commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>