build(justfile): Fix "just prune-old-compile-env" target
The target doesn't work because there's a "|| true" in the middle of the
command that prevents the execution of "docker image rm": the command is
interpreted as:

    docker image list ... | grep -v ... || (true | xargs docker rmi ...)

so the xargs command never runs. Let's fix this. The "|| true" is not
needed anyway because "grep -v" does not fail if it doesn't find the
pattern to exclude.

Signed-off-by: Quentin Monnet <qmo@qmon.net>