feat(nat): Make allocator return Option-al IP addresses
The IP/port allocator may, or may not return NAT-related target source
and destination addresses and ports, depending on whether NAT is
desirable for the considered connection. For ports, we already used an
Option<NatPort>; make the allocator use one for the IP addresses, too.

Adjust the stateful NAT code accordingly, depending on whether we need
to NAT or not (based on whether we get Some IP address, or None).

Signed-off-by: Quentin Monnet <qmo@qmon.net>