feat(nat): Add proper error handling in stateful NAT ports manipulation
Make set_source_port() and set_destination_port() correctly handle and
pass up any error that might happen when trying to update the source or
destination ports, respectively, as part of the NAT processing for a
packet.

Also change them to take a non-optional NatPort object as an argument
(unwrap the Option in the caller rather than in methods that shouldn't
be called at all if there's no port to translate).

Signed-off-by: Quentin Monnet <qmo@qmon.net>