feat(routing): reapply config on frr restart
Detect within the router io loop that frr has been restarted and
re-apply the last config if any. The detection method is somewhat
hacky as it just checks that the number of connection requests
has changed. However, this avoids needing to propagate rio or the
frrmi object up to the rpc handlers, which unnecessarily
complicates the code.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>