feat(config): transfer mgmt models to config
* separate external/underlay
* unifies error codes
* splits external and internal config

This crate does not define how configurations are to be stored.
It only defines the types to represent and validate configurations.
A top-level GwConfig object is the entity that keeps a complete
snapshot of the configuration. For convenience, it contains an
GwConfigMeta object to keep metadata set by the management.

The organization of this crate is:

├── display.rs   // impl Display for some types in this crate
├── errors.rs    // contains single error type for the whole crate
├── external     // model for the external/api-facing config.
│   ├── mod.rs
│   ├── overlay
│   └── underlay
├── gwconfig.rs // defines GwConfig object
├── internal    // internal configuration model
│   ├── device
│   ├── interfaces
│   ├── mod.rs
│   ├── natconfig
│   └── routing
└── lib.rs

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>