feat(nat): Rework allocator structure, add allocator to StatefulNat
Rework our allocator draft to introduce a NatAllocator trait and a
default NatDefaultAllocator implementation. It should be easy to replace
this default allocator with alternative NatAllocator implementations for
future experiments.

The default allocator implementation is currently only a skeleton, and
will be completed in future commits.

One change worth noting is that we no longer "find" the NAT pool prior
to calling the allocation function. Instead, the allocator itself will
be responsible for looking up from the right pool - or whatever internal
structure it decides to use - and return a valid IP for the Peering in
use, based on the information it gets from the NatTuple object.

Signed-off-by: Quentin Monnet <qmo@qmon.net>