feat(nat): Implement port allocation for stateful NAT
Implement port allocation for stateful NAT. Introduce structures
PortBlockAllocator, PortAllocator.

The idea is the following: we split the L4 ports space in "blocks" of
256 ports. A PortAllocator works on one of these blocks, and allocates
ports until the block is full. A PortBlockAllocator is a collection of
all the blocks forming the 65k-port range, listed in a randomised order,
and holds states (free, heating (being allocated), or cooling (fully
allocated, waiting for connections timeouts to expire before being
garbage-collected)) for each of these blocks. It also contains a
PortAllocator object, to work with the block currently in use for
allocating new ports. Ports are allocated in a linear fashion within a
port block.

Signed-off-by: Quentin Monnet <qmo@qmon.net>