feat(nat): Introduce stateful NAT IP allocator
Introduce an IP allocator for stateful NAT. This is a simple IP
allocator relying on a large, albeit compressed, bitmap. With up to 2^32
elements, it can map the whole IPv4 address range. For IPv6, we can map
it to 2^32 elements, which should realistically cover the use cases we
can afford: with 65k ports per IP address, this makes for about 277
trillion connections, each requiring multiple bytes in memory.

Allocation for IPv4 and IPv6 differs a little: our bitmap is large
enough to cover the IPv4 space, so we can just translate IPv4 addresses
from and to u32 integers. For IPv6, the bitmap does not represent all
possible addresses, but instead we use it to store up to the first 4^32
addresses from our list of allowed prefixes for the VpcExpose in use,
which means we have a small conversion step between the bitmap lookup
and the obtention of the address to allocate.

Signed-off-by: Quentin Monnet <qmo@qmon.net>