feat(nat): Implement the default allocator for stateful NAT
Based on the IP and port allocator submodules that we implemented in the
last commits, connect the different parts to implement the default
IP/port allocator for stateful NAT.

This default allocator contains several maps of NatPool objects, using
VPD-related IDs, destination addresses, and L4 protocol type to form
keys. For each packet, the ID, L4 protocol, and destination address, all
passed as part of the packet tuple, are used to retrieve the two NatPool
objects to use for source and destination NAT, respectively. If there is
no NatPool for a given direction, then NAT is not required for this
direction, for the connection under consideration.

Signed-off-by: Quentin Monnet <qmo@qmon.net>