feat(nat): Use dst_vni metadata to simplify stateless NAT
Now that we can rely on a dedicated pipeline stage to perform the
destination discriminant lookup (at the moment, destination VNI), we do
not need to keep this indirection level in the NAT tables to find the
rule (virtually) associated with the destination VNI for source NAT.

Instead of keeping the two lookups - first find the index of the
NatPrefixRuleTable in the NatPeerRuleTable, then consult the
NatPrefixRuleTable - simplify the structure of the NAT tables for source
NAT, to only contain a per-destination-VNI hashmap containing the rules.
This makes the NAT code simpler, in particular for building the rules
tables.

Adjust unit tests to manually set the dst_vni metadata when relevant.

Signed-off-by: Quentin Monnet <qmo@qmon.net>