refactor(nat): Make add_peering() a method for PerVniTable
When adding add_peering, we didn't want to make it part of the
implementation of PerVniTable: even if it makes sense to have it a
method of PerVniTable, we preferred to keep it in a file separate from
the rest of the struct implementation, because this function requires on
definitions from our external configuration models, unlike the rest of
the implementation.

Guess what? We don't actually have to move it to the same file if we
want to make the method part of PerVniTable! It still makes sense to
make add_peering() a method for PerVniTable today, so let's do just
that.

No functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>