feat(routing): use PrefixMapTrie wo/ default
Use PrefixMapTrie instead of PrefixMapTrieWithDefault in fib and
rib. Both already handle the fact that there will always be a
"default" route. Most importantly, the change is needed because
using the version with default forces us to implement the Default
trait on the types, which we will not be able to implement for
FibRoutes. The rib is also changed to PrefixMapTrie for homogeneity
and because there are some common utils shared by both fib/rib
in display.rs.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>