feat(routing): delete vrfs on stale timeout
It may happen that, while FRR is restarting, some config change
happens such that a VRF is no longer needed. All of the routes
in that VRF will eventually be removed, since FRR will not
refresh them, but we need to make sure that the VRF gets removed
too since we only remove vrfs once FRR deletes the routes, and in
that case it won't. So, after flushing stale routes, check if
VRFs can actually be removed.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>