feat(routing): no mark as stale the default route
Add a method to check if a route is the "default preset drop"
route that each VRF has on creation, and prevent it from being
marked as stale since FRR will never refresh it.
Nothing bad would happen if we did because such a route cannot
be deleted anyway: attempting to delete it (e.g. because it was
marked as stale) would just replace it with the same route not
marked as stale.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>