ci: Run dependency updates weekly rather than daily
It's great to remain on top of dependency updates. It's not so great to
spend time reviewing (and triggering CI) for update PRs every day or so.
It should be enough to run the workflow for updates once a week, rather
than every day - especially now that updates from the dpdk-sys
repository are no longer synchronised via this workflow.

Signed-off-by: Quentin Monnet <qmo@qmon.net>