feat(prefix): Extend trait IpPrefix with a last_address() method
Add a last_address() method to trait IpPrefix, and implement it for
Ipv4Prefix and Ipv6Prefix. This method returns the last IP address from
a given prefix. Under the hood, it returns the .broadcast() address, but
in our context the last address of a prefix is not necessarily a
broadcast address, so we named the method differently.

This is in prevision for future work for stateful NAT.

Signed-off-by: Quentin Monnet <qmo@qmon.net>