feat(prefix): Add size() method to IpPrefix trait
We have a size() method for our struct Prefix, but we don't have it for
the IPv4/IPv6 "subtypes" that implement the IpPrefix trait. Add the
method to that trait, implement it for IPv4 and IPv6, and make the
implementation for Prefix rely on those for Ipv4Prefix/Ipv6Prefix.

This is so we can obtain the size of an Ipv4Prefix or Ipv6Prefix object
in future work for stateful NAT.

Signed-off-by: Quentin Monnet <qmo@qmon.net>