feat(routing): Use RefCell instead of UnsafeCell
The prior implementation of FibGroupStore was using &*fg on
an UnsafeCell to manipulate and fetch FibGroups in the
FibGroupStore.

A better approach is to just use RefCell which uses a flag
to dynamically do the proper borrow checking to make sure that we
do not mutate a FibGroup when someone else is actively holding
a reference to it.

In this commit, we move the Fib to use an Rc<RefCell<FibGroup>>
and pass around Ref<FibGroup> and Ref<FibEntry> to make sure the
compiler knows when there is an outstanding reference to a
FibGroup. The RefCell guarantees dynamically that no one else
has a reference to the FibGroup when we modify it.

We guarantee that access to the FibGroup in a store is exclusive
during modification by ensuring that all Rc's to a FibGroup are
protected by the Fib's left_right::ReadGuard or left_right::WriteGuard.
Modifications to the FibGroups are protected by left_right::WriteGuard.

This commit now marks the method that modifies the FibGroups as
unsafe so that we can ensure that we properly check that we satisfy
the correct invariants when calling those methods.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>