fix(routing): Remove possibly incorrect use of unsafe
The vrf table code was getting a reference to a VRF
in the by_id table and then calling by_id.values_mut
which in theory can invalidate the original reference.
That this does not happen now does not mean it won't
happen in the future.

The fix is to remove the unsafe code and use the
values_mut iterator to find the default VRF.  This has
some performance drawbacks, but that is fine for now.
A better fix that is sound, does not use unsafe, and is
performant would be welcome.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>