feat(routing): speed up values_mut_except_default
The previous implementation of values_mut_except_default()
replaced a simpler yet unsafe method to get mutable references
to all vrfs in a vrf table and another one for the default vrf.
However, it required two passes over the collection.
This patch changes values_mut_except_default() back to use unsafe
to provide a mutable iterator for all vrfs except the default
and an immutable -potentially unsafe- reference to the default vrf.
The advantage is that the method now returns an iterator instead
of a vector.

The safety of the reference to the default vrf is granted because:
  * just creating an iterator should not change the underlying
    collection, even if it is a mutable one.
  * one might argue that using the iterator may cause such a
    reference to become invalid. This could happen if the
    underlying collection (hashmap) changed the address of the
    default vrf. This could happen if new element were inserted/
    removed from the hashmap. But that would require having a
    mutable reference to it, which the borrow-checker will forbid
    while the mutable iterator is in scope.
  * the default vrf cannot be deleted
  * regardless of the above, to be extra safe, we Box the vrfs so
    that even if the elements within the collection changed their
    memory locations while holding that reference, the addresses of
    objects wouldn't. Boxing should suffice because, in principle
    the only possible move would be putting the vrf on the stack,
    which the borrow checker would forbid again. So, pinning
    should not be needed.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>