feat(config): augment Underlay object with vtep
Currently, we learn about the vtep from a vtep-type interface
specified in the underlay vrf, as per our current api. This creates
significant confusion in the current code as we try to recover
that information when building the internal configuration, which
requires the creation of intermediate types an significant noise.

To simplify this, augment the Underlay object with an optional
VtepConfig object pretending that it was specified like this in the
api. This allows validating the configuration more easily and
building the internal config without the need for intermediate
conversions since we can just copy the vtep config into the
internal one, as done in the next commit.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>