feat(config): further simplify vtep config
Simplify the code creating the vtep config from an interface
config and remove validation from the interfaces since we validate
the vtep already.

The code may be further simplified by moving the TryFrom impl.
added in this commit to the converters so that the rest of the
code already sees the vtep configuration.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>