fix(ci): Avoid pulling debug compile-env just for "cargo deny check"
Pass the profile argument for the current job to the "just" invocation
at the "cargo deny check" step of CI runs. Without this argument, the
step would default to the "debug" profile and would _always_ pull the
debug compile-env image, even for jobs where all other steps use the
"release" profile, adding a useless image pull. Now that we pass the
argument, the step should use the same image as the rest of the job,
either "debug" or "release" depending on which job matrix case we run.

Fixes: e0dc86f54151 ("test(ci): factor sterile.yml into dev.yml")
Signed-off-by: Quentin Monnet <qmo@qmon.net>