fix(concurrency-macros): Fix shuttle/loom macro when all features are on
When both "shuttle" and "silence_clippy" are turned on in the
concurrency crate, then the "with_shuttle" macro is never defined. If
another crate attempts to use #[concurrency_mode(shuttle)] in that case,
the code fails to compile due to the missing macro. The same happens
for the "with_loom" macro, under similar conditions.

Fix this by defining the "empty" version of the macros whenever
"silence_clippy" is on.

Fixes: fcd72a029250 ("feat(concurrency,concurrency_macros): Add conditional compile macros")
Signed-off-by: Quentin Monnet <qmo@qmon.net>