refactor(nat): Rework state creation from allocation in stateful NAT
This is... Well, not sure what happens here exactly. This is a
transition commit created while re-organising the Pull Request, and it's
here to minimise the changes in future commits. We reorganise the way we
manipulate the allocated resources (IP and port) for stateful NAT
sessions, and how we create the flow state from them - although state
creation itself is left for a future commit.

Just move on.

Signed-off-by: Quentin Monnet <qmo@qmon.net>