feat(nat): Introduce struct for allocated IP and port for stateful NAT
Rather than just independent IP address and port, introduce a
AllocatedIpPort<I: NatIp> struct to hold the result from the allocation.
This container is not really defined yet in this commit, we simply have
a placeholder implementation to make it work. The real implementation
will be part of the actual allocator and will come in a follow-up
commit.

Having the definition of the AllocatedIpPort allows us to implement
function new_state_from_alloc(), where we build the state for creating a
flow table entry based on the resources returned by the allocator.

Signed-off-by: Quentin Monnet <qmo@qmon.net>