chore(nat): Change NatPort type for stateful NAT
We use NatPort as a checked type to hold L4 port values for stateful
NAT. We made sure ports would only be > 1024, to avoid allocating a
reserved port. But this is too restrictive, as the allocator may need to
assign reserved ports, for example for source NAT on the return path:

    forward path: (a:2000, B:80) -> (A:3000, b:7000)
    reverse path: (b:7000, A:3000) -> (B:80, a:2000)
                                        ^
        we need to reserve port 80 for source NAT for the reverse flow

Change the NatPort type: Make it a non-zero u16 instead (value 0 is not
a valid port value).

Signed-off-by: Quentin Monnet <qmo@qmon.net>