feat(nat): Add port allocation to stateful NAT allocator
Implement the PortAllocator object contained in an AllocatedIp object,
and the AllocatedPort object returned at the end of the allocation
process.

The PortAllocator contains a slice of AllocatorPortBlock objects with
the state of each 256-port "block" of ports for the given IP. It also
contains an AllocatedPortBlockMap, containing references to blocks
(AllocatedPortBlock) in use. These blocks of ports allow distinct
threads to work each on a separate chunk of the L4 port space, avoiding
contention. At last, the AllocatedPortBlock contains a Bitmap256 to
manage the 256 ports it represents, and can allocate AllocatedPort
objects.

Port blocks are allocated in a "random" order (the order is randomised
when building the PortAllocator object), in an attempt to randomise port
allocation.

Signed-off-by: Quentin Monnet <qmo@qmon.net>