feat(nat): Implement deallocation of resources for stateful NAT
When an allocated object, namely a struct AllocatedPort representing a
port and a reference to an IP address, gets out of scope and dropped, we
want to deallocate the ports and (possibly) the IP that this object
represents.

To do so, we implement the Drop trait and override the drop() functions
for the relevant objects. This allows use to establish a concurrent
access model where we don't need to hold too many locks to be able to
deallocate resources: we work with weak references in lists pointing to
allocated IPs and ports, but we gave back references to make an
AllocatedIp point to its parent AllocatedPortBlock, the
AllocatedPortBlock point to its parent AllocatedIp, and the AllocatedIp
point to the IpAllocator. When the AllocatedPort object gets dropped,
all weak references become invalid, ensuring other threads won't pick up
the allocated resources, and we can then deallocate the relevant
resources from the drop() function, accessing the parent allocator of
each object via the back references.

Now that our allocator, with its deallocation path, is architecturally
complete, we can add a diagram to help visualise the relationships
between the different components.

Signed-off-by: Quentin Monnet <qmo@qmon.net>