feat(nat): Add allocation for reverse-path resources to NAT allocator
When we get a packet from a new flow, we want to allocate resources (Ip
address and port) for source and destination NAT for this packet. But
this is not all: we also need to "reserve" the required resources from
the allocator for the return path for the flow. If we don't, the
corresponding resources might get allocated for a different flow,
causing collisions and potential issues.

Make sure our allocator can reserve the resources for the return path
when allocating for a new flow, and return these resources alongside
the result for the path forward.

Signed-off-by: Quentin Monnet <qmo@qmon.net>