feat(nat): Make stateful NAT pipeline stage use the default allocator
Two things happen in this commit. First, we make the stateful NAT
pipeline stage use our newly impmented default, bitmap-based IP and port
allocator.

As a consequence, a second thing happens: we also have to compute the
tuple for the return path for the flow, so that a corresponding entry
can be added to session table.

We also remove the placeholders in nat/src/stateful/allocator.rs now
that we have our real implementations, they're not longer needed.

Signed-off-by: Quentin Monnet <qmo@qmon.net>