feat(nat): Add code to build the stateful NAT allocator from a VpcTable
Similar to what we have for stateless NAT, add the code to build the
context - in our case, the IP and port allocator - for stateful NAT.

Note: This code is not called yet! We'll need to call it, likely from
the mgmt crate, in the future.

Now that the relevant functions are used for building up the allocator,
we can remove the clippy warning about dead code for the stateful NAT
module - we just need to keep it for the NAT sessions submodule.

Signed-off-by: Quentin Monnet <qmo@qmon.net>