feat(nat): Move stateless NAT prefix size validation out of "config"
As part of the validation step for all VpcExpose objects in a VpcTable,
we ensure that, for each expose containing prefixes for NAT, the number
of IP addresses defined by the initial list of prefixes is equal to the
number of IP addresses defined by the target list of prefixes. This is
to ensure that we can apply a 1:1 mapping between IP addresses of each
set.

This validation step is essential for stateless NAT, however, we cannot
apply it indistinctively when processing the Gateway config, because it
would prevent users from having IP sets of different sizes, for example
if they're using stateful NAT rather than stateless NAT.

As a first step towards a more flexible approach, move the validation
step from the validation functions for the VpcTable in the config crate
to the nat crate, and call it from mgmt right before building the
stateless NAT tables.

A follow-up step will be required to conditionally validate the VpcTable
for stateless NAT and build the NAT tables in mgmt, once we have an API
to indicate whether or not stateless NAT is desired.

Signed-off-by: Quentin Monnet <qmo@qmon.net>