style: Replace tabs with spaces
We use speaces nearly everywhere for indentation in the project, meaning
most of us have our editors configured to use spaces, accordingly. As a
result, each time I add a new crate in the toplevel Cargo.toml, I need
to rework my change to convert the spaces to tabs for consistency...

Let's just clean up these few tabs that are here for no apparent reason,
and use spaces instead. That's just three files to modify, two of which
had inconsistent indent, with that.

Signed-off-by: Quentin Monnet <qmo@qmon.net>