feat(nat): Build new stateful NAT allocator when processing config
This commit links the code from the NatAllocatorWriter to the building
blocks from the default allocator to construct a new allocator when
processing a configuration. It only deals with the case when we don't
have an existing allocator and build a new one from scratch; updating an
existing allocator is not implemented yet.

Note that the NatAllocatorWriter is not generic, and directly calls the
functions from the NatDefaultAllocator. We could make it generic, by
extending the NatAllocator trait, at the expense of passing the generic
argument in many locations and exposing the NatDefaultAllocator higher
up in the mgmt code, which brings little benefit. If we want to test new
allocators, it remains easy to swap the function called in
build_new_allocator() (and, in the future, in function
update_existing_allocator()).

Signed-off-by: Quentin Monnet <qmo@qmon.net>