fix(nat): Use separate allocators for TCP and UDP
We use sepearate entries in the NatPool table for TCP and UDP
allocators, however, these separate entries have been using Arc
references pointing to the same objects (with the same bitmaps). As a
consequence, we've been using a single allocator for both TCP and UDP
allocations for any given VpcExpose. This is not what we want: TCP and
UDP should be decorrelated when it comes to allocating IP addresses and
ports.

Fix this by using distinct allocators for TCP and UDP. We will likely
have to use yet another one for ICMP in the future.

Fixes: fae5f7c66d9e ("feat(nat): Add code to build the stateful NAT allocator from a VpcTable")
Signed-off-by: Quentin Monnet <qmo@qmon.net>