feat(nat): Pass allocator from config to StatefulNat stage
Introduce a "reader" for the stateful NAT allocator such that instead of
owning the allocator directly, the StatefulNat pipeline stage owns a
reader pointing to the allocator that a NatAllocatorWriter can write
into. This allows the mgmt crate to make new allocators, generated from
user-provided configurations, visible to the StatefulNat stage for
processing packets.

Signed-off-by: Quentin Monnet <qmo@qmon.net>