feat(nat): Update stateful NAT allocator for new user config
When processing a new configuration, replace the existing NAT allocator
with a new one.

See comments: longer term, we'd prefer to report the internal state of
the old allocator into the new one, to keep existing connections open,
but we don't support that at the moment. Also, we need to adjust the
existing entries in the flow table accordingly - right now, we need to
drop them all - but this will wait until we switch stateful NAT to the
new flow table implementation.

Signed-off-by: Quentin Monnet <qmo@qmon.net>