feat(pipeline,others): make pipelines replicable
Packets are processed according to a pipeline, currently by a
single thread. In order to allow several threads (e.g. DPDK
workers) to process packets according to the same pipeline, we
need pipelines (and their stages) to be somehow extended. The
option of sharing pipelines/stages may not always be viable or
convenient: some stages have no config, some have read-only config
and some may need to keep state that we may want to be shared
or be distinct per thread to avoid locking penalties.

This patch extends the pipeline crate and its traits so that
pipelines can be replicated such that, when multiple packet workers
exist, each can use one dedicated replica. It also implements the
replication logic for all of the existing network functions.

The capability of building replicas of network functions and
pipelines seems to be the best approach to the problem because:
  * most stages that keep some config / read-only state do so by
    keeping a reference or device to an external piece of state
    (e.g. a left-right reader). Hence, replicating the stage
    is nearly zero-cost in terms of memory footprint as only
    the "reader" object to access the state is duplicated.
  * avoids the need for locking pipelines.
  * allows keeping independent (fully-disjoint) stages (e.g.
    those that would keep dynamic per-thread state).
  * The way how a given network function or "stage" is replicated
    cannot be anticipated / assumed by the pipeline crate. There-
    fore the replication logic is part of the network function
    definition. This gives significant flexibility to accommodate
    many cases and, while it requires the implementation of a
    method per network function, such implementations for the
    network functions currently supported is very simple.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>