feat(mgmt,dataplane): Make NAT context optional
When we build a pipeline for the dataplane, we may, or may not, want to
use stateless NAT. Similarly, we may or may not want stateful NAT. We
will likely have use cases for which we don't want both NAT stages.

Currently, the code in mgmt and dataplane builds the context for both
pipeline stages anyway, even though the corresponding stages are not in
the pipeline. Other than wasting resources, it may enforce unnecessary
limitations. For example, stateless NAT requires that users specify
lists of prefixes containing the same number of available addresses in
the private and public sets of addresses (before and after NAT), because
it relies on a 1:1 mapping of the IP addresses between the two sets.
This restriction is enforced when we apply a configuration to build the
context for stateless NAT: if we build this context when it is not
required, we end up applying unconditionally the same restriction for
stateful NAT, too!

Instead of that, let's make it so the NAT context objects can be
optional. This way, we don't need to create them - and to enforce their
respective additional validation steps - if they are not necessary for
the pipeline in use.

We may have a pipeline with both stateless and stateful NAT stages. At
the moment, both operations would apply to all packets: we do not have a
way to determine from the user configuration what "expose" objects
should use stateless, or stateful, NAT. In the future, we'll need to
find a way to support this, but for now, we'll work with the limitation
on the sizes of IP addresses sets if we use both modes - using two NAT
operations on each packet doesn't really make sense at the moment
anyway.

Note that objects for other pipeline stages, such as the flow table, for
example, could be made optional as well. This commit only focuses on
NAT-related context objects, the other may be addressed in the future.

Signed-off-by: Quentin Monnet <qmo@qmon.net>