feat(dataplane): Move to a pipeline factory for multithreading
Previously, setup_pipeline was passed as a method that could
only be called once.  But this means that the driver modules
cannot create a pipeline per thread.  Moreover, pipeline stages
are not cloneable either, nor are they Send, so they couldn't
be sent to remove threads.

Now, setup_pipeline is a Send + Sync closure that is passed
to the drivers. With that change, the driver can create threads
and send setup_pipeline to those threads, and then each worker
can call setup_pipeline to create its own private copy of the
pipeline.  Moreover, pipeline stages do not have to be Send or
Sync, just the stuff that builds the stage.  And the compiler will
tell you if you get this wrong.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>