feat(nat): Transition stateful NAT to new flow table implementation
Make the stateful NAT pipeline stage switch to the new flow table
implementation. This includes:

- Changing the .sessions attribute to point to an Arc<FlowTable>
- Updating the relevant steps accordingly: tuple extraction (now
  consisting in collecting information from the prior flow table lookup
  stage), session lookup, and sessions creation.

As for the tests, those that rely on flow table lookups need an
adjustment to make sure that we effectively collect the information (we
need to mimic the NAT-related portion of the flow table lookup stage;
thankfully, it's very simple to do).

This commit leaves many code areas in the StatefulNat implementation to
be improved, now that we've switched to the new flow table: we can get
rid of some portions of code, we can also simplify some steps due to the
flow table now working with IpAddr (rather than internally embedding two
tables, for Ipv4Addr and Ipv6Addr, like the previous one). We need to
rework how we use NatTuple vs FlowKey, and to use the results from the
allocation (rather than NatState objects) as values for session entries,
to ensure proper garbage collection in the allocator. All these steps
will come in future work.

Signed-off-by: Quentin Monnet <qmo@qmon.net>