ci: Use GitHub App token for bump workflow
We have a GitHub App set up for the dataplane repository, so that
updating the dpdk-sys repository can automatically create a Pull Request
for the dataplane repository to update the changes to the compile-env
image. We should be able to reuse this App when creating Pull Requests
to bump dependencies in the dataplane repository: the expected advantage
is that the generated Pull Requests should now trigger CI on creation,
without the need for reviewers to close and re-open the Pull Request to
manually triggre it.

Signed-off-by: Quentin Monnet <qmo@qmon.net>