refactor(nat): Use FlowKey instead of NatTuple in stateful NAT code
We used to rely on type NatTuple to represent the information of an
incoming packet: source and destination IP addresses, protocol and
ports. Now that we have switched to the new flow table implementation,
all this information is available from the FlowKey structure, and there
is no need to have two different objects to hold this.

This commit removes all instances of NatTuple objects from the code, and
replace them with FlowKey objects. As the NatTuple were used all over
the stateful NAT logic and the allocator code, this means a lot of
replacements, and a large commit.

One difference worth noting between the retired NatTuple and the FlowKey
objects is that FlowKey uses IpAddr addresses, whereas NatTuple would be
generic for types implementing NatIp (in practice, Ipv4Addr/Ipv6Addr).
We could propagate this difference all the way through, which would
result in changing the structure of the allocator table and working with
IpAddr addresses only. Instead, we stop the propagation at the
"entrance" of the allocator code, and convert them with
NatIp::try_from_addr().

The tests also need adjustements to turn the manually-declared NatTuple
objects into FlowKey objects. Test test_tuple_extraction() has been
removed, given that extraction itself (parsing the packet) is no longer
part of the NAT code. The rest of the tests remain unchanged.

Signed-off-by: Quentin Monnet <qmo@qmon.net>