refactor(nat): Use VpcDiscriminant instead of NatVpcId in stateful NAT
Type NatVpcId, wrapping around a Vni object, was introduced to work with
VNIs - the only discriminants supported when we introduced the stateful
NAT code. They are still the only discriminants supported, but now we
have a better wrapper VpcDiscriminant, an enum that may represent
alternative values in the future.

This commit replaces the use of NatVpcId with VpcDiscriminant. To make
sure we don't handle unsupported objects by mistake, the allocator makes
sure that the discriminants are actually of type VNI for the packet.
The code that builds the allocator context also explicitly uses VNIs
when building the PoolTable entries.

Note that this refactor is "naive", with the objective to simplify the
code and make it more future-proof: it does not focus on resources
usage. Storing the bigger wrapper VpcDiscriminant over the NatVpcId
may use more memory for each PoolTable entry (it looks bigger as an
enum, although the compiler might be able to make it the same size, I
haven't checked).

Signed-off-by: Quentin Monnet <qmo@qmon.net>