refactor(nat): Make stateful_translate() IP version-agnostic
Move the complexity for handling IP versions from stateful_translate()
in the stateful NAT code to the net crate, as methods try_set_source()
and try_set_destination() in the Net object. The resulting code in
stateful_translate() is much easier to follow.

Also update net's utils to avoid duplicating the code.

Signed-off-by: Quentin Monnet <qmo@qmon.net>