test: introduce vm test runner
Introducing the vm test runner.  Thanks to the shiny new testn crate(s),
we are able to run tests in micro vms.  This is very handy for test
cases where we need to do privileged operations.

Sadly, this doesn't play nicly with the extant test-runner.sh script.
So, for the moment, dataplane is excluded from the containerized runner.
This is not a loss really, since

1. no other tests exist for that package at the moment,
2. even if they did, containers are not sufficient to test that crate.
3. the vm test runner uses a container to start the vm anyway.

A follow on commit will be required to make CI compatible with this
change, as we are currently missing hugepage allocation.  You also
(currently) need to manually pull a container image to make the vm
runner function.  I'm not 100% sure how to side step that requirement.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>