feat(net,pkt-meta): Add basic ICMPv4 support in FlowKey
Add an IcmpProtoKey struct and use it with IpProtoKey::Icmp(icmp).

We only add support for ICMPv4 at this time.

The new key type is not trivial to add, because ICMP messages can come
in a variety of flavours, depending on their types. Some are "ICMP Query
Messages", and have an ID that we'll need to update for NAT; others are
"ICMP Error Messages", and embed a fragment of the original IPv4 packet
header that triggered them, and we'll need to update the addresses and
ports in this fragment. For the latter case, we introduce an
EmbeddedPacketData type which looks like a light version of FlowKeyData,
but we can't use FlowKeyData because it would make a recursive type
definition.

This is a large commit because updating the enum value for IpProtoKey
means we get to update the bolero tests as well, and we need some
methods in struct Icmp4 to update these correctly.

Note that these methods are not all implemented yet, as we're missing
the parser code for accessing the inner IPv4 packet. This is left for
follow-up work.

Signed-off-by: Quentin Monnet <qmo@qmon.net>