feat(tracectl): refactor for usability
If many targets exist, each with multiple tags, it can be difficult
to know which tag needs to be used to control a given target.
Ease this by explicitly associating a name to each target and
automatically promoting it as a tag so that it needs not be
specified as a tag.

Also, keep state to know if a target is custom or not and use a
distinct macro (custom_target!) to easily spot in the code if
a target is custom or not.

Finally, calling the controller registry methods from the outside
of the crate is not needed since we can register third- party crate
targets with the custom_target! macro. So, update the README.md
explaining that option and make the register method private.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>