chore(dataplane): Transmit Box<Packet<_>> to avoid copies
Previously, we were passing Packet<_> explicitly on the worker channels
in the kernel driver.  That could lead to moves which will copy the
packet meta data, etc. during send, which is bad.  This code explicitly
uses Box<Packet<_>> instead.  It also moves to using a functional style
with iterators for some of the code to make the tx to workers a bit more
idiomatic Rust.

Signed-off-by: Manish Vachharajani <manish@githedgehog.com>