feat(tracectl,mdataplane): config as string
Add a method to show the tracong configuration as a string and
add --tracing-config-generate cmd line option to show it.

Note: the config string provided is the config that would be
used to configure the dataplane using tags; it does not correspond
to the actual config internally passed to envfilter with target
real paths/ids.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>