feat(tracectl): fix random test failures
The tests in tracectl are tricky because tracing registration
happens at link time. Therefore, the tracing database contains
targets declared across distinct test functions. This was making
some tests fail depending on the execution order of the tests.
Some of the tests will be fixed in subsequent PRs.

Commiting this to avoid pipeline failures.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>