feat: ignore compile-env symlink
If you are on NixOS (as I am now), then the existing workflow based
on fake nix is broken.  I have a hack workaround for now (although I
want to contribute a better solution soon).  My hack involves making
compile-env a symlink.  However, the current .gitignore assumes that
compile-env is a directory.  This breaks many rebases and is
constantly in my way.

The easy fix is to add compile-env to .gitignore.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>