feat(tracectl): add tracing db reset method
Rationale: currently, it's not been decided how external components
(e.g. gateway agent) will learn about existing targets & tags. As
a result, the API is non-exhaustive in that it may include the
log-levels of target/tag to be changed, but not all of them. Those
not included will not be changed. So far so good. However, that
also means that if the loglevel for a tag is removed from the
config, the system will not update it since it only attempts to
change the levels that have been explicitly specified.

Solve this by resetting the log-levels to the initial setting.
With this, the semantics of the tracing configuration are: "use
the default values and override those explicitly stated in this
configuration".

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>