feat(tracectl): serialize tests
The tracing control implementation relies on a static tracing
controller. As a result, that shared controller can be accessed
by multiple tests if run concurrently. This makes it very difficult
to test and was causing random-looking failures. Fix this by
annotating all tests bt #[serial] from the serial_test crate.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>