feat(net): Introduce EmbeddedHeaders to parse ICMP's inner packets
Introduce new enum EmbeddedHeader and struct EmbeddedHeaders,
counterparts to enum Header and struct Headers, but for parsing IP
packets embedded inside of ICMP Error messages (such as "Destination
Unreachable").

It is not possible to reuse the existing enum and structs for IP packets
embedded in ICMP messages, because the packets may not be complete, but
we still need to access the transport-layer ports if available, for
example, even if the transport header is truncated; and the
implementation of the Parse trait for the existing structs does not
allow that (nor should it).

Make an (optional) EmbeddedHeaders object part of struct Headers, so we
can hold the parsed data and reuse it when "deparsing" the packet,
although the DeParse trait implementation for EmbeddedHeaders will be
completed in a follow-up commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>