feat(net): Add TruncatedUdp, TruncatedTcp header types
TruncatedTcp is an enum that can be either a regular TCP header, or a
"truncated TCP header", meaning that it contains at least 8 bytes - the
source and destination header of the original TCP header - plus an
undetermined number of additional bytes of header, possiby 0. This
truncated header can be used to represent the beginning of a TCP header
in an IP packet embedded as payload for ICMP Error messages. We need to
be able to access to the TCP ports for this embedded packet for NAT,
because we need to update them for NAT.

TruncatedUdp, as one may have guessed, implements the same structure for
possibly-truncated UDP headers.

Signed-off-by: Quentin Monnet <qmo@qmon.net>