feat(net): Implement ParsePayload for enum EmbeddedHeader
Now that we have our new struct EmbeddedHeaders, implement parsing for
the payloads of the various IP and related headers towards enum
EmbeddedHeader objects. The new parse_embedded_payload() come with a bit
of duplcation from the existing parse_payload(), but it's not that bad.
They're not entirely the same: they use TruncatedTcp and TruncatedUdp
instead of Tcp/Udp, and they only process header types that can be
present in the embedded IP packet.

Once we have these methods in place, we can complete the implementation
of the ParsePayload trait for enum EmbeddedHeader.

Signed-off-by: Quentin Monnet <qmo@qmon.net>