feat(tracectl,mngmt,config): handle removals
The tracing controller applies changes to an existing config.
However, the tracing configuration in the API is optional and
non-exhaustive. We need to handle the case where some tracing
configuration exists and a new configuration is received that
does not contain any tracing configuration. Currently, the
loglevels of the prior configuration would be used. However,
a more correct behavior should be to setting the loglevels back
to their defaults.

Implement this by providing a default TracingConfig to the
tracing controller in case the API configuration does not
include one. The controller is already applying some presets
to handle the case of a partial configuration.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>