feat(id): rkyv id
The dataplane may not (yet) have the ability to dynamically allocate memory
at early startup if we intend to use DPDK's allocator.

This will require us to use a zero copy deserialization mechanism to understand
the config we give it.

Id is planned to be used in that config, so add zero copy deserialization support
(rkyv) to that crate.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>