feat(net): Implement incremental checksum update
Implement incremental checksum update for IP, TCP, UDP 1s' complement
checksums, following the algorithm described in RFC 1624, along with
accompanying unit tests.

Contrarily to regular checksum update, we cannot rely on etherparse
here, as the crate doesn't support incremental updates at the moment, so
we provide them as a default method for the trait. This requires making
the Checksum associated type convertible to and from a 16, which is
already the case for the types used in all existing trait
implementations anyway.

This will be useful to update the checksums of an inner IP packet
embedded in an ICMP error message, for example, after updating addresses
and/or ports for this inner packet.

Signed-off-by: Quentin Monnet <qmo@qmon.net>