feat(init): introduce InitErr
Introduce a package wide application error type.

This is very different from our more typical library level error types
in that it is intended to be a catch-all error type for fatal errors.

Almost nothing which can go wrong in this crate is expected to be
recoverable without human intervention.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>