chore(net): Remove unused ParsePayload trait
Commit 6913affe0df6 ("refactor(net): Remove ParsePayload implementation
for specific headers") removed the ParsePayload implementation of many
Header variants, because the trait was never used. It left the existing
implementation for Header, and then we implemented it for EmbeddedHeader
in commit e13503ec12f6 ("feat(net): Implement ParsePayload for enum
EmbeddedHeader"). But the trait remains, in fact, unused. Remove it and
make parse_payload() simple methods for the enums.

Signed-off-by: Quentin Monnet <qmo@qmon.net>