refactor(net): Allow checksum handling to return errors
At the moment, methods from the Checksum trait never fail (they never
return an error). In the future, however, we plan to implement the trait
for some additional header types where looking up, computing, or setting
a checksum may fail. This can be because the header is truncated or
immutable, for example. To prepare for this, change the signature for
the methods to make them return an Option<> or a Result<>.

When updating the callers for the checksum-handling methods, we do rely
on the fact that for most header types, checksum manipulation does not
file, and we allow ourselves to unwrap the return value considering the
error case is unreachable.

Signed-off-by: Quentin Monnet <qmo@qmon.net>