feat(net): Update embedded IP's checksum in Headers' update_checksum()
When updating checksums for a packet, we update the nework-layer and
transport-layer headers' checksums for the packet. If this packet is an
ICMP Error message and embeds a (possibly truncated) IP packet, we want
to update the checksum for that inner IP header, too. Let's add it.

As mentioned in the comment, it would be desirable to update the inner
transport layer as well. But we refrain from doing that, because it's
possible that the inner packet is truncated, in which case we cannot
compute the new checksum correctly. We can check whether the payload is
full and update accordingly, or we can perform an unconditional,
incremental update of the checksum, based on the old and new values
changed in the header.

Given that the unconditional, incremental update looks cheaper (and more
robust) than evaluating whether the inner packet is full, we choose this
option. This means we cannot do the incremental checksum update at this
location, because we do not know what fields have changed in the header
and what their old values were. Instead, leave it to the processing
code, such as the NAT logic, to run the transport checksum update as
they see fit, and skip the layer here in update_checksum().

Signed-off-by: Quentin Monnet <qmo@qmon.net>