feat(net): Store inner packet's payload length in check_full_payload()
If we find out that a possibly-truncated IP packet embedded in an ICMP
Error message is, in fact, full, it may be useful to store its actual
size or transport-layer payload size, to be able to reuse it later
without having to recompute all lengths to account for the possible ICMP
Extension Structures at the end of the packet. Here we modify
EmbeddedHeaders's full_payload to turn it from a boolean to an
Option<u16>, to store the length.

Signed-off-by: Quentin Monnet <qmo@qmon.net>