feat(net): Implement IcmpAny wrapper around Icmp4 and Icmp6
Implement a new IcmpAny header type that wraps around either Icmp4 or
Icmp6. The two protocols have differences, but they also share
similarities, in particular for handling embedded IP packets in ICMP
Error messages. This wrapper will be useful to process such packets for
NAT in follow-up work.

Annoyingly, we can't really take the Icmp4 or Icmp6 object out of its
parent Headers to wrap it into an IcmpAny object instead, so we've got
to use references. But these references are either mutable or immutable,
and we'd like to support both cases: so we get two variant of the
object, IcmpAny and IcmpAnyMut, because I couldn't find a better
alternative.

We implement the Checksum type for the wrappers, although in some cases
we return errors, such as when trying to update the checksum for the
immutable variant, or when the wrong payload type is passed.

Signed-off-by: Quentin Monnet <qmo@qmon.net>