refactor(nat): Move NatPort out of stateful NAT module
NatPort is only used in stateful NAT at the moment, but it's not tied to
its internals and could just as well be used for stateless port
translation in the future. Move it out of the stateful NAT module.

Along with it, move NatTranslationData to the root of the module, so we
can use it in a follow-up commits for methods shared between stateful
and stateless NAT.

Signed-off-by: Quentin Monnet <qmo@qmon.net>