feat(left-right-tlcache): avoid potential race
Augment get_factory() trait method to return also the identity and
version for a given key. This avoids the potential race that could
happen due to the fact that the provider is queried twice on each
access to an rhandle and nothing prevents the provider contents
from changing between the two.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>