test(net): Add Bolero generators for ICMP Error message inner headers
In order to parse the inner IP packets potentially truncated,
potentially embedded in an ICMP Error message, we rely on the
EmbeddedHeaders structure. This commit adds Bolero generators to produce
instances of this struct for use with ICMP Error messages. The
generators are not used yet; this is in preparation for tests related to
future work in the nat crate (ICMP's inner packets translation).

Signed-off-by: Quentin Monnet <qmo@qmon.net>