test(net): Add Bolero generators for ICMP Extension Structure objects
Implement an abstract representation of ICMP Extension Structure (for
ICMP v4 and v6), and arrays to store them. Implement the DeParse trait
in order to get the size of these arrays, and to write them to a packet
buffer. Implement a Bolero generator for these arrays of Extension
Structure trailers.

This gives us the means to test NAT processing of ICMP Error messages'
inner packets, when these messages contain ICMP Extension Structures, in
follow-up work.

Signed-off-by: Quentin Monnet <qmo@qmon.net>