test(net): Add a Bolero generator for a full ICMP Error message packet
Add a Bolero generator to produce valid ICMP Error message datagrams,
including:

- Ethernet header
- Outer IP header (v4/v6)
- ICMP header (v4/v6)
- Optional inner IP header (v4/v6)
- Optional inner transport header, possibly truncated (TCP/UDP)
- Optional payload
- Optional padding and ICMP Extension Structure trailer objects

As one might expect, the optional payload depends on the inner transport
header being present and full, which in turns depends on the inner
network header being present. ICMP extensions, however, do not require
the previous field to be present or complete.

We set the lengths for the different headers:

- For the outer IP layer, set the payload length as expected
- For ICMP header, set the optional length only if we use ICMP extensions
- For the inner IP layer, set the length to the actual inner payload
  (assumed full) or to a theoretical higher value (payload assumed
  truncated).
- For the inner transport layer, if UDP, set the length to the same
  theoretical value (ajusted for the transport layer)

This is in order to test processing of such ICMP Error messages, in
particular in the context of NAT, where some fields of the inner headers
may require translation.

Signed-off-by: Quentin Monnet <qmo@qmon.net>