fix(deps): pin dplane-rpc to commit SHA instead of version
Version isn't considered in git dependencies, so this isn't doing what we think it
is doing.

Instead, it is just following the default branch for the repo while pretending to be pinned.

This should be resolved by cutting a proper tag for that repo, but in the mean time we can pin
to a commit.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>