ci: Remove Claude workflow execution on Issues assignment
It's not really clear why Claude should run on assignments in addition
to comments and Issue creation. And then, whose permissions should we
check when assigning? The Issue author? In that case it means we've
already run the workflow at issue creation if it was possible. The
assignee? The assigner? We can't check their "author_association" from
the API, so that would make things tricky.

I also fear this may create too much noise if we create the issue with a
mention in the description, then assign the issue. We'll have two
interventions from Claude in that case.

Let's remove execution from assignments. If it is desirable to get
Claude's perspective after assigning, it remains possible to mention the
agent directly in a comment.

Signed-off-by: Quentin Monnet <qmo@qmon.net>