feat(routing): add version to FibTable
Versioning the fibtable will allow us to determine the validity
of cached readhandles without needing to ask the provider, which
would add an extra lookup (e.g. hash or similar), on every
cache lookup. This patch adds the version for the FibTable and
the logic to monotonically increment its value anytime it is
updated. Note that version is u64 instead of AtomicU64 since
we don't need to synchronize multiple threads on its value.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>