feat(routing): implement fibtable cache
Declare thread-local ReadHandleCache `FIBTABLE_CACHE`.
Implement trait ReadHandleProvider for FibTable so that it can
be polled by the cache, per thread.

We could implement ReadHandleProvider for FibTableReader instead.
However, implementing it for the FibTable is more general in that
it would also work if the FibTable itself was not wrapped in left-
right.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>