fix(net): Pass the right buffer slice when parsing Icmp4/Icmp6
We do not account for the consumed bytes in the packet's buffer when
passing the slice to Icmp4.parse_with()/Icmp6.parse_with(), usually
resulting in a failure to parse the headers. Let's adjust by using the
right offset: the length of the buffer, minus the length remaining to
parse.

Fixes: 803338d3667d ("feat(net): Parse payload for ICMP")
Signed-off-by: Quentin Monnet <qmo@qmon.net>