feat(left-right-tlcache): optimization for iter
The existing iterator implementation is subobtimal in that each
call to get_iter() needs to request the provider a full iterator
for the whole set of read handle factories, unless refresh arg is
set to false, and then refresh the local cache with read handles
built from each factory. This is inefficient in case the set
of objects in the provider rarely changes (which may be the case
in most setups).

Avoid pulling an iterator from the provider if the version did
not change, as this means that no objects have been added or
removed. If anything, we remove unusable readers.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>