feat(left-right-tlcache): minimize num of rhandles
Fix the code so that a cache keeps the minimal number of rhandles
after a refresh. The previous code would keep an independent
rhandle for aliases, meaning that if there are N objects each with
one allias, 2 x N rhandles would exist after a refresh.
Solve this by letting aliases use the same rhandle as their
primary.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>