chore(net): Add placeholder for passing IcmpAny's checksum errors
Say we define a new error enum for a header or packet, and this enum has
a variant for checksum-related errors. When we return that variant, we
probably encountered a CheckSum error somewhere in the code, and it may
be worth to "embed" it inside of the new error variant. But
ChecksumError is a generic type, and takes a generic parameter which is
an object implementing the Checksum type.

Concretely, this means that if we want to pass ChecksumError objects up
from IcmpAny<'_> or IcmpAnyMut<'_>, we need to pass a
ChecksumError<IcmpAny<'_>> - note the lifetime. This lifetime will in
turn propagate to the new error type, and to every location using this
error type. For the intended use case, for example, this would mean
adding a lifetime parameter to all functions handling StatefulNatErrors,
which is both useless and painful.

Instead, use a "hack" to get rid of the lifetime: introduce a
placeholder struct (IcmpAnyChecksumErrorPlaceholder), and convert from
ChecksumError<IcmpAny<'_>> into an error with that struct. It's a bit
dirty, but it does the job.

Signed-off-by: Quentin Monnet <qmo@qmon.net>