feat(dataplane): Add stateful NAT stage to pipeline
Now that users have an API to specify whether they desire stateful or
stateless NAT, add a stateful NAT pipeline stage to dataplane's
pipeline, to support processing packets with, well, stateful NAT,
wheneverrequired.

To pass the reader to the DynPipeline closure, we create a
NatAllocatorReaderFactory object which is not really necessary (we could
just get the reader and clone() it), but helps make the code more
consistent with stateless NAT's setup.

Signed-off-by: Quentin Monnet <qmo@qmon.net>