chore(dataplane,mgmt): Make NAT context writers mandatory
This reverts changes made in commit 27577d48ae92 ("feat(mgmt,dataplane):
Make NAT context optional"), where we made these objects optional. It
was necessary at the time to avoid validation conflicts due to the
different validation requirements for stateful and stateless NAT: if one
context object was optional, we could just skip validation for it.

Now that we've introduced a way to determine, through the user API,
whether stateful or stateless NAT is required for every VpcExpose
object, we no longer need to have the writers optional.

Signed-off-by: Quentin Monnet <qmo@qmon.net>