test: Make VpcExpose in NAT tests use stateful or stateless NAT
We now use VpcExposeNat objects to tell whether a VpcExpose should use
stateless or stateful NAT (or none). We don't actually use this
parameter yet, but we will in a future commit. As a preparation step,
make VpcExpose objects in NAT tests use stateless or stateful NAT,
depending on the tests context.

Signed-off-by: Quentin Monnet <qmo@qmon.net>