feat(dataplane): add single pkt-io to pipeline
Add a pkt-io instance to pipeline before egress stage. The injection
buffer is 10K packets. The punting buffer is only 100 packets large.
This is temporary: we currently mark some packets as local,
but do not handle them yet. Therefore, the first 100 of local
packets will be cached (forever) and the rest removed.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>