fix(init): break up init step 1/3; split out sysfs
I thought I could keep this functionality specific to the init crate,
but I almost immediately needed it in the dataplane itself during
startup.

As such, this is the first of three commits which are intended to split
the init package up and refactor it to function more like a library
than was originally envisioned.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>