feat(net): Support ICMP-in-ICMP
So we recently added support for parsing ICMP Error messages' inner
messages. We processed the network layer of the inner packet, and the
transport layer of the inner packet, which can be only TCP or UDP.

... Says who? As it turns out, we can just as well have an error message
for an ICMP Echo Request, for example, in which case we need to NAT the
message just like for TCP or UDP.

Add the required infrastructure. It's more code that I'd like, but I
don't really see a simple alternative: we get new traits for
EmbeddedTransport, and support for potentially truncated ICMP headers
messages, even though it's unlikely we get truncated ones - RFC 792 says
there should be at least 8 bytes of the transport header, which is
enough to hold most ICMPv4 or ICMPv6 headers (as defined in etherparse,
at least), with the exception of Timestamp Requests and Replies. The
code for the truncated headers (TruncatedIcmp4, TruncatedIcmp6) is
largely derived from TruncatedTcp and TruncatedUdp. Similarly, we
implement Checksum for these new types. The annoying bit, as always when
adding ICMP support for a transport layer, is that we suddenly get a
header that may not have source and destination ports. Happily, the only
location where this matters is in NAT code, where we don't support fully
ICMP yet - we'll add the relevant processing soon.

Signed-off-by: Quentin Monnet <qmo@qmon.net>