feat(interface-manager): provide TapDevice
The existing implementation was not building a TapDevice object
from the file descriptor bound to the kernel interface, needed to
read/write to the tap. Fix this by letting TapDevice::open()
return a TapDevice, even if that is not needed by the interface
manager. Since taps are persisted, the expectation is that opening
again the tap from elsewhere will yield a new TapDevice with a
usable descriptor.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>