fix(nat): Fix outer IP header stateful NAT for ICMP Error messages
To process ICMP Error messages in NAT, and in particular in stateful
NAT, we don't need to process the packet like a regular one with mapping
from the flow session table, even after translating the inner packet.
Instead, we need to follow RFC 5508 (see comments in the code, in the
commit).

This means that:

- We shouldn't expect to find a mapping for the packet, we can get ICMP
  Error messages processing out of the "hot path".
- We shouldn't rely on regular translation for the outer header, either.
- Instead, move processing to the ICMP-dedicated method, that we rename
  with a more suitable name.

Notes: the fix is not complete. We need to address properly the case
where the ICMP Error packet comes from the private realm, but we're
missing some parts from that:

- We need a per-IP mapping, which we don't have at the moment (we want
  this in the future to be compliant with REQ-1 from both RFCs 4787 and
  5382).
- We need to know the public IP address of the gateway, for the case
  where there is no mapping to use, and we don't know this in the NAT
  crate at this time.

Fixes: fab39efe0735 ("feat(nat): Translate ICMP Error messages' inner IP packet")
Signed-off-by: Quentin Monnet <qmo@qmon.net>