feat(nat): Change NatPort to allow (possibly-null) identifiers
NatPort has been used in NAT to represent ports that cannot be zero (the
value is invalid for TCP and UDP). Now that we work with ICMP as well,
and that we need to translate an identifier that can be null, we need to
rework the structure. Change it to an enum, so we can keep the validated
type for TCP and UDP. This will be used in follow-up commits to extend
support for ICMP in the NAT pipeline stage.

Signed-off-by: Quentin Monnet <qmo@qmon.net>