feat(net): define trait for buffer pool
We need to abstract the entity to allocate packet buffers so that
the packet injection logic can allocate both TestBuffers and DPDK
Mbufs. This commit defines the bare minimum abstraction and
constraints that we require for a packet buffer pool.

This commit defines the trait and implements it for TestBuffers.
The implementation of DPDK is todo!()

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>