feat(interface-manager): make TapDevice friendlier
We need to read and write asynchronously from/to TapDevices.
Such operations have to be concurrent: writes cannot block reads
or viceversa. Therefore, they need to be implemented in separate
tokio tasks. However, the existing methods to read/write from
TapDevice require &mut TapDevice, making this hard.

Possible solutions:

1) We could use a Tokio Mutex, to provide interior mutability.
However, this easily leads to deadlocks: we need to acquire the
lock to read/write but when attempting those operations, control
is given back to runtime, without us being able to release the lock:
when we acquire the lock, we don't know if the subsequent read/write
will "block" or not.

2) Provide owned copies of the file descriptors to distinct tasks.
This approach works, but is fragile and easy to break when using
tap devices, because stale descriptors will easily prevent us from
reopening tap devices, causing resource-busy errors.

3) Change the implementation of TapDevice so that read and writes
do not require &mut self and can therefore be shared by distinct
tasks (e.g. within an Arc).

This patch provides (3), given that 2) is significantly more complex.
This is achieved as follows:
  - tokio::fs::File is replaced by std::fs::File
  - the underlying descriptors are set as non-blocking
  - descriptors are wrapped by AsyncFd so that we can test if they
    are writable/readable.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>