test(nat): Fix Bolero tests for ICMP embedded in ICMP Error messages
The Bolero tests for parsing ICMP Error messages' inner packets were
written with TCP and UDP in mind. You know, _ports_. But now we want to
support ICMP as well, with its lone identifier. We need a number of
adjustments to the test.

As part of that we also add several methods to ICMP objects to check
whether they're ICMP "Query" messages (Echo, Timestamp) and to retrieve
their identifier. These methods will be reused in the processing logic
in follow-up commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>