refactor(nat): Refactor identical functions translate_inner_icmp{4,6}()
The functions for translating the inner ICMP header embedded within an
ICMP Error message are the same for ICMP v4 and v6, the only thing
preventing us to merge them so far is that they take different types of
headers as arguments. Refactor these functions into a generic one, using
the existing Checksum trait as well as a new TruncatedIcmpAny trait to
make generic the two additional, identifier-related methods we need.

Signed-off-by: Quentin Monnet <qmo@qmon.net>