feat(nat): Reject stateful NAT connection if there is no source NAT rule
In the absence of a valid rule for source stateful NAT, as derived from
the absence of entry in the NatPool object for source NAT, then we
should reject the new connection. This is to avoid endpoints from a VPC
to reach the outside if NAT is set up, but there is no corresponding
entry for NAT-ing a packet so it can go out. Note that we block if
there's no source NAT rule, and regardless of destination NAT rules:
this is because we do not want to allow external hosts to initiate
connections.

Signed-off-by: Quentin Monnet <qmo@qmon.net>