fix(nat): Do not try to set destination port for ICMP
For stateful NAT, when applying the translation, we process TCP/UDP and
ICMP differently for the source port: TCP/UDP get their source ports
updated, ICMP gets its identifier updated. For the destination port, we
only process TCP/UDP: we've done the ICMP identifier already, there's
nothing more to do.

Doing so, we introduced a bug, because we try to set the destination
port for ICMP anyway. Of course, it fails, and causes the function to
return an error. This still results in the packet being translated, but
because of the error being returned, the packet is marked as "done" and
will be dropped rather than forwarded.

Let's add pattern matching to ensure we only try to set destination port
for TCP/UDP, not for ICMP.

Fixes: d51311bd12ce ("feat(nat): Add logic for processing ICMP in stateful NAT")
Signed-off-by: Quentin Monnet <qmo@qmon.net>