feat(nat): Use more precise reasons for packet drops (stateless NAT)
We need more precision when dropping a packet from stateless NAT than
just "NatFailure". We have a range of existing reasons, let's map them
to the errors returned from the NAT logic - just like we did for
stateful NAT in a previous commit.

One drawback is that it may not be as explicit as before that the error
occurred at the NAT stage.

Signed-off-by: Quentin Monnet <qmo@qmon.net>