fix(net): Support ICMP as embedded transport layer for flow keys
When we added support for ICMP as a transport layer for ICMP Error
messages' inner packets, we forgot to update some portions of the code.
In particular, enum InnerIpProtoKey, used to represent the transport
layer of this embedded packet, supports TCP and UDP, but not ICMP!

So in this commit, we add the new variant. And we add the code for
building this new variant; which means extending ICMP-related structs
with a few utility functions. And, behold! It turned out we missed some
other parts. So we also add the code to build the
IcmpProtoKey::ErrorMsgData variant from a packet.

We also update the related tests accordingly, including some Bolero
generators. Some of the new code is not actually exercised (see FIXME:
comment in the code), because the generator in use for the related test
never produces embedded headers for ICMP Error messages - this commit
does not go as far as to add that.

At the end of the day that's a lot of changes, but all related to ICMP
support in flow keys.

This went under the radar because:
- We don't have tests for ICMP Error messages processing with stateful
  NAT yet (one is coming in a follow-up commit).
- Existing Bolero tests do not exercise all of the code, due to
  limitations in the generator in use (we never produce embedded headers
  for ICMP Error messages in this test).

Fixes: fd6047791901 ("fix(net): ICMP-in-ICMP for NAT: try harder")
Signed-off-by: Quentin Monnet <qmo@qmon.net>