feat(args): completely rework args crate
The args crate is now has a wider scope: it is not just responsible
for parsing and validating arguments, it is (partly) responsible for
handing the dataplane a memfd file with parsed and acted on data.

The idea here is to

1. allow all of the highly sensitive, highely privileged, and one off jobs
   to be done by dataplane-init rather than dataplane
2. so that dataplane can init the DPDK eal as quickly as possible after process
   start.

The memfd strategy (developed in this PR), is a relatively simple way to accomplish
these objectives (although there is a lot of room for improvement here).

Signed-off-by: Daniel Noland <daniel@githedgehog.com>