feat(dataplane): drivers call portmap
Let packet drivers call build_portmap() (or the async version),
to create taps for ports, and provide a writer to the portmap
table, already populated with the port-to-interface mappings.

Drivers temporarily own the table and its writer, and may
create readers from a factory if needed, but they should
return the writer.

 NOTE: this commit only implements that functionality for the
      kernel driver.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>